rename migrations
